Jupyter Notebook of main file
Was thinking we could make the main python file into a notebook and all of the steps in the mean time (would be a good report version of the code). 
When it comes to the data processing itself, I deleted the columns: 'timestamp_created', 'timestamp_updated', 'comment_count', "steam_china_location". 
Only kept the rows where language == english.